backend/pyproject.toml
[tool.poetry]
name = "numma-exchange"
version = "0.1.0"
description = "aoaoaoaoa"
authors = ["Axel791 <HS.Kazakov@mail.ru>"]
license = "BSD-3"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.6"
django-split-settings = "^1.2.0"
pillow = "^9.5.0"
celery = "^5.2.7"
redis = "^4.5.4"
gunicorn = "^20.1.0"
djangorestframework = "^3.14.0"
aiogram = "^2.25.1"
pytelegrambotapi = "^4.11.0"
loguru = "^0.7.0"
amocrm-api = "^2.5.0"
django-cors-headers = "^4.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
